✨ feat [#11] pr수정사항 적용 (colorFrameSize enum 별도파일로 분리)
