+++ src/components/chips/color-frame.tsx
@@ -1,6 +1,6 @@
 import styles from "./color-frame.module.css";
 import { CHIP_COLORS, type ChipColor } from "@/constants/chips/chip-colors";
-import { ColorFrameSize } from "@/constants/common/common-size";
+import { ColorFrameSize } from "@/constants/chips/color-frame-size";
 import { ButtonHTMLAttributes } from "react";
 
 interface ColorFrameProps extends ButtonHTMLAttributes<HTMLButtonElement> {
