+++ src/constants/chips/color-frame-size.ts
@@ -1,0 +1,6 @@
+export enum ColorFrameSize {
+  XSmall = "xsmall",
+  Small = "small",
+  Medium = "medium",
+  Large = "large",
+}
